requirements.txt
python-dotenv==1.0.0
requests>=2.26.0
urllib3==1.26.6
pyrebase4==4.5.0
requests-toolbelt==0.9.1
email-validator==2.1.0
python-jose==3.3.0

# tiktoken==0.5.1
chromadb==0.5.21
langchain==0.3.9
openai==1.56.2

pydantic[email]>=2.7.4
setuptools==80.3.0
fastapi==0.109.0
uvicorn==0.27.0
firebase-admin==6.2.0